refactoring: move institutionGroupBaseHandler fn
