+++ apps/backoffice/src/app/api/institutions/[institutionId]/groups/any/route.ts
@@ -1,6 +1,9 @@
-import { institutionGroupBaseHandler } from "@/lib/be/groups/institution-groups-util";
-import { BackOfficeUserEnriched, withJWTAuthHandler } from "@/lib/be/wrappers";
-import { NextRequest, NextResponse } from "next/server";
+import {
+  type BackOfficeUserEnriched,
+  withJWTAuthHandler,
+} from "@/lib/be/wrappers";
+import { type NextRequest, NextResponse } from "next/server";
+import { institutionGroupBaseHandler } from "../handler";
 
 /**
  * @operationId checkInstitutionGroupsExistence
