+++ apps/backoffice/src/app/api/institutions/[institutionId]/groups/route.ts
@@ -1,10 +1,11 @@
-import {
-  institutionGroupBaseHandler,
-  toGroupsResponse,
-} from "@/lib/be/groups/institution-groups-util";
+import { toGroupsResponse } from "@/lib/be/groups/institution-groups-util";
 import { sanitizedNextResponseJson } from "@/lib/be/sanitize";
-import { BackOfficeUserEnriched, withJWTAuthHandler } from "@/lib/be/wrappers";
-import { NextRequest } from "next/server";
+import {
+  type BackOfficeUserEnriched,
+  withJWTAuthHandler,
+} from "@/lib/be/wrappers";
+import type { NextRequest } from "next/server";
+import { institutionGroupBaseHandler } from "./handler";
 
 /**
  * @operationId getInstitutionGroups
